move current_simulation to job.py, resolve the class in one place
- current_simulation() now lives next to get_job_context in job.py, which
  is where the job-context accessors belong.
- get_inference_headers() takes the caller's configured inference_class
  and returns the resolved header, so one function owns the precedence.
  inference.LLM hands its class over and no longer sets the priority
  header itself.
- the existing text-simulation log line in AgentSession.start now names
  the inference class too, so the demotion is visible at session start.
